hello worlding and notes exploring electron.
I'm not vendoring the whole toolchain into git because it's huge,
but you can see where I put it in the launch script.

It's fortunately pretty easy to download the electron zip from
their github releases.  (The webpage directs you to do things that
are a thousand times more complicated for no reason, of course.)