is asking for just types about electron any less fucked?  no, in fact it's exactly the same outcome.
i... don't understand this.  The electron package page on npmjs.org
actually directly mentions this "it'll return a path if you use it in
node" outcome, but... okay?
What the fuck am I supposed to *do* about it...?
That is the least helpful and useful behavior I've ever seen, and this
_should_ be such a simple thing that it's driving me absolutely bonk.