okay, watch is a little tricky.  and paths are fun.  and the electron app doesn't actually pick up changes, either, still.
but, um, okay.
Let's take a next step of "make sure we can have preload and renderer
scripts" and do that in typescript before we go any further.

Also I should probably check that things like filesystem and proc exec
are actually viable in this bun/node mangle of a world,
because that's... not yet proven, and kinda important.