okay, renderer demo success (with js, not ts), and it can read stuff from preload.  cool.
apparently relative paths in html files is fine, and it's resolved
according to... where that html file is on disk.  So the path to
*get to it* was already settled in the loadFile call in main code.
Alright, that works for me.

(Not sure it'll be the same if we start doing custom protocol stuff
and having any kind of internal router layers.)