trying to figure out how to have a typed API between renderer and main.
Not... actually sure it's possible, frankly.

The electron ipc parts kinda ditch all info.
And also just mechanically, I still don't know how to import anything
or really use typescript at all in the preload file, which is where
we'd want to sanity check the interfaces.  (I think.)

So that's not coming together like I want :/

But also: here's how you yell at typescript to assume the type for
something unknown, demonstrated on that versions object.  And that
appears to be working just fine and dandy, so that's neat.