add api description to a subdir.  mostly so we don't bother to transpile that file.
At the moment, my build instruction is... everything named "*.ts"
in the src dir.  I don't feel like changing that to an enumeration.

Something different is likely enough in the future.  For now this
satisfices and keeps the dist dir clear of silliness.

The current behavior of bun is to bundle the content of zow into the
generated renderer.js file, which is is clear and fine.