oh.  okay, node_modules management works... less like I want it to than... uh.
Apparently it runs fine if it searches up and finds what it already needs.

And it runs fine even if there's no bun lockfile, because, lol, lmao.

but if I run `bun install` in this dir (mostly to generate the
lockfile honestly), that... drops a new hot node_modules in the dir.

This is... okay, I don't want to talk about it.